[CIFS] Minor cleanup to EA patch
CC: Jeff Layton <jlayton@redhat.com>

+++ fs/cifs/cifssmb.c
@@ -5393,7 +5393,7 @@ QAllEAsRetry:
 	temp_fea = ea_response_data->list;
 	temp_ptr = (char *)temp_fea;
 	while (list_len > 0) {
-		__u8 name_len;
+		int name_len;
 		__u16 value_len;
 
 		list_len -= 4;
